+++ CHANGES.rst
@@ -7,6 +7,7 @@ mio 0.1.6.dev
 
 - Added support for packaging Python `wheels <http://pythonwheels.com/>`_.
 - Define a default ``init()`` method on ``Object`` and just make ``TCloneable init()`` just call ``init()``.
+- Added ``String strip()`` method.
 
 
 mio 0.1.5 (2013-11-28)